// SwsPN/ClientApp/src/app/core/models/errors.ts
export class Errors {
    swsPn: string;   // __EMPTY_4
    swsName?: string;   // __EMPTY_5
    partType?: string;  // __EMPTY_7
    classFlag1?: string; // __EMPTY_8
    classFlag2?: string; // __EMPTY_9
    classFlag3?: string; // __EMPTY_10
    classFlag4?: string; // __EMPTY_11
    classFlag5?: string; // __EMPTY_12
    classFlag6?: string; // __EMPTY_13
    tempMh?: string; // __EMPTY_14
    operationNum?: string;  // From JH1 File
    prdNum?: string;
    prdLevel?: string;
    areaC?: string;
    custC?: string;
    checkCode?: string;
    winUserID?: string;
    dateError?: Date;
    urlDirectory?: string;
    reported?: boolean;
    dateReported?: Date;
    drawing?: string
  }
  